+++ tests/testthat/test-interaction.R
@@ -9,6 +9,15 @@ test_that(": in interactions become x", {
                       update=FALSE)
 })
 
+test_that("fixest i() becomes =", {
+  testthat::skip_if_not_installed("fixest")
+  library(fixest)
+  mod <- feols(Ozone ~ Solar.R + i(Month), airquality)
+  expect_known_output(modelsummary(mod, "markdown", gof_map = list()),
+                      file="known_output/msummary_fixest_i.md",
+                      print=TRUE,
+                      update=FALSE)
+})
 
 test_that("conditional conversion of : to x", {
     mod <- lm(am ~ drat : mpg, mtcars)
@@ -17,7 +26,7 @@ test_that("conditional conversion of : to x", {
         output = "dataframe",
         coef_rename = c("DRAT" = "drat"))
     expect_true("drat:mpg" %in% tab$term)
-                      
+
     mod <- lm(mpg ~ disp, mtcars)
     tab <- modelsummary(
         mod,